perl-compatible extensions lists in config.sh
There are tests that read config.sh and expect module lists
in the Configure format.

To work around this, there are two copies for some of the lists
now: one (dynamic_ext) perl-compatible for configpm and test and
whatever else may need it, and another one (fullpath_dynamic_ext)
with proper directory names for use in Makefiles.

Since we're back to the perl format now, configpm patch can be
dropped for all perl versions.