don't add "all" target to module builds
When building with --all-static, this would previously try to build
both the "all" and "static" targets, which would fail if the toolchain
could only produce static libraries/executables.  Here, I've removed
the "all" target from all the module building lines, for symmetry.
Those that don't specify a target will still default to all because
MakeMaker ensures that in the Makefiles it generates.